% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{columbus_crime}
\alias{columbus_crime}
\title{Crime Rates in Columbus}
\format{A data frame with 49 rows and 3 variables:
\describe{
  \item{crime_i}{Indexed crime rate.}
  \item{income}{household income in thousands.}
  \item{home_value}{Home values in thousands.}
}}
\source{
Anselin, Luc. 1988. Spatial econometrics: methods and models.
  Dordrecht: Kluwer Academic, Table 12.1 p. 189.
}
\usage{
columbus_crime
}
\description{
A datset containing information on crime rates, home values, and median
incomes for neighborhoods in Columbus, Ohio.
}
\note{
This data is obtained using \link[spdep]{oldcol}
}
\keyword{datasets}

