% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppois.R
\name{lagsarpois.filoglik}
\alias{lagsarpois.filoglik}
\title{Full-information maximum likelihood function for spatial poisson}
\usage{
lagsarpois.filoglik(theta, y, X, W)
}
\arguments{
\item{theta}{Parameter vector consisting of rho, and any betas.}

\item{y}{n-length vector of dependent values}

\item{X}{$n\times p$ matrix of independent covariates, created by model.matrix}

\item{W}{A CsparseMatrix of the weights objects.}
}
\value{
A numeric calculation of the model likelihood.
}
\description{
Full-information maximum likelihood function for spatial poisson
}

