% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppois.R
\name{lagsarpois}
\alias{lagsarpois}
\title{Spatial Autoregressive Poisson Model}
\usage{
lagsarpois(formula, data = list(), listw, type = c("lag", "mixed"))
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. The details of
model specification are given for \code{\link[stats]{lm}}}

\item{listw}{A \code{\link[spdep]{listw}} object created for example by
\code{\link[spdep]{nb2listw}}}

\item{type}{Default "lag", may be set to "mixed" for a spatial Durbin model;
when "mixed", the lagged intercept is dropped for spatial weights style
"W", that is row-standardised weights, but otherwise included. \em{mixed not
currently supported.}}

\item{dat}{An optional data frame containing the variables in the model.  By
default the variables are taken from the environment which the function  is
called.}
}
\description{
This function implements a limited-information maximum likelihood estimator
for Poisson regression models. The estimator was described by
\href{http://dx.doi.org/10.1016/j.regsciurbeco.2010.04.001}{Lambert,
Brown, and Florax (2010)}.
}
\examples{
  lagsarpois(counts ~ outcome + treatment, data = dd, listw = W)

# from ?glm
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
# random spatial distance, symmetric and zero diagonal
relationship <- matrix(runif(81, 0, 1), 9)
relationship[lower.tri(relationship)] <- t(relationship)[lower.tri(relationship)]
diag(relationship) <- 0
W <- spdep::mat2listw(relationship)

}
\seealso{
\code{\link[spdep]{lagsarlm}}
}

