% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarpoisson.R
\name{sarpois.filoglik}
\alias{sarpois.filoglik}
\title{Full-information maximum likelihood function for spatial poisson}
\usage{
sarpois.filoglik(p, y, X, W)
}
\arguments{
\item{p}{Parameter vector consisting of rho, and any betas.}

\item{y}{n-length vector of dependent values}

\item{X}{\eqn{n \times p} matrix of independent covariates, created by
model.matrix}

\item{W}{A CsparseMatrix representation of the weights objects.}
}
\value{
A numeric calculation of the model likelihood.
}
\description{
Full-information maximum likelihood function for spatial poisson
}

