% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarpoisson.R
\name{sarpoisson}
\alias{sarpoisson}
\title{Spatial Autoregressive Poisson Model}
\usage{
sarpoisson(formula, data = list(), listw = NULL, method = "liml", ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. The details of
model specification are given for \code{\link[stats]{lm}}}

\item{data}{An optional data frame containing the variables in the model.  By
default the variables are taken from the environment which the function  is
called.}

\item{listw}{A listw object created for example by \code{\link[spdep]{nb2listw}}}

\item{method}{The method to be used for fitting the regression equation.
Defaults to "liml", a limited-information maximum likelihood. Other
options are "fiml" (full-information maximum likelihood), "model.matrix" to
return a model matrix, and "non-spatial", which will execute a non-spatial
Poisson regression (identical to \code{\link[stats]{glm}}).}

\item{...}{Further arguments passed to \link[stats]{nlm}.}
}
\value{
A list of class `sarpoisson` containing the following components:
\describe{
  \item{coefficients}{The estimated coefficient values.}
  \item{fitted.values}{The estimated mean of the poisson distribution.}
  \item{residuals}{Difference between estimated mean and observed value.}
  \item{df.residuals}{Degrees of freedom remaining in residuals.}
  \item{df.null}{Degrees of freedom in the null model.}
  \item{logLik}{Numerical log likelihood.}
  \item{rank}{Rank of the model.}
  \item{call}{Model estimation call.}
  \item{nlm_results}{Complete results from \link[stats]{nlm}.}
  \item{information.matrix}{Fischer information matrix, obtained as inverse
  of Hessian.}
  \item{method}{The method used in maximum likelihood estimation.}
  \item{xlevels}{}
  \item{terms}{Terms of the model frame}
}
}
\description{
This function implements a limited-information maximum likelihood estimator
for Poisson regression models. The estimator was described by
\href{http://dx.doi.org/10.1016/j.regsciurbeco.2010.04.001}{Lambert,
Brown, and Florax (2010)}.
}
\examples{
  summary(sarpoisson(crime_i ~ income + home_value, data = columbus_crime,
          method = "non-spatial"))
  summary(sarpoisson(crime_i ~ income + home_value, data = columbus_crime,
          listw = columbus_neighbors, method = "fiml"))

}
\seealso{
\code{\link[spdep]{lagsarlm}}
}
