% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarpoisson.R
\name{set_sarpoisson}
\alias{set_sarpoisson}
\title{Build a glm class object from an NLM poisson fit}
\usage{
set_sarpoisson(object, mf, mt)
}
\arguments{
\item{object}{An object containing the estimation results from an NLM
optimization routine.}

\item{mf}{The model frame.}

\item{mt}{The model terms.}
}
\value{
An object of class `c("glm", "lm")`
}
\description{
Build a glm class object from an NLM poisson fit
}

