% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppois.R
\docType{package}
\name{sppois}
\alias{sppois}
\title{sppois: A package for estimating spatial poisson regressions.}
\description{
The spppois package provides functions for estimating and summarizing output
from spatial autoregressive Poisson count models. The primary
estimation fuction is \link{sarpoisson}
}
