% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarpoisson_methods.R
\name{summary.sarpoisson}
\alias{summary.sarpoisson}
\title{Summarizing SAR Poisson Models}
\usage{
\method{summary}{sarpoisson}(object)
}
\arguments{
\item{object}{An object of class `sarpoisson`, usually a result of a call to
\link{sarpoisson}}
}
\description{
`summary` method for class "`sarpoisson`"
}

