% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarpoisson-tidiers.R
\name{tidy.sarpoisson}
\alias{tidy.sarpoisson}
\alias{sppois_tidiers}
\title{Tidying methods for spatially autoregressive poisson models}
\usage{
\method{tidy}{sarpoisson}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{An object returned from [sppois::sarpoisson()]}
}
\description{
These methods tidy the coefficients of spatial autoregression
models generated by functions in the `spatialreg` package.
}
\examples{
\dontrun{
library(sppois)
crime_sppois <- sarpoisson(crime_i ~ income + home_value, data = columbus_crime,
                           listw = columbus_neighbors, method = "fiml")

tidy(crime_sppois)
tidy(crime_sppois, conf.int = TRUE, conf.level = 0.99)
glance(crime_sppois)

}
}
