% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wqueen}
\alias{wqueen}
\title{Queen-style county adjacency matrix}
\format{
An object of class \code{dgCMatrix} with 3078 rows and 3078 columns.
}
\usage{
wqueen
}
\description{
A \code{CsparseMatrix} representation of county contiguity in the United States.
}
\keyword{datasets}
